requirements.txt
azure-functions
requests
types-requests
azure-cosmos
python-dotenv